Update inside outside activity detail
